fix: update package imports to satisfy linter
The way Docusaurus aliases some imports makes it impossible for the
linter to find the source. It looks like this is a side-effect of their
transition from JS to TS (could be somethign else though). This might be
a temporary change as type support increases and/or the beta version
progresses.